+++ apps/meteor/client/views/room/contextualBar/Discussions/DiscussionsListRow.tsx
@@ -1,6 +1,7 @@
 import type { IDiscussionMessage } from '@rocket.chat/core-typings';
 import type { MouseEvent } from 'react';
 import { memo } from 'react';
+import { useTranslation } from 'react-i18next';
 
 import DiscussionsListItem from './components/DiscussionsListItem';
 import { useTimeAgo } from '../../../../hooks/useTimeAgo';
@@ -13,9 +14,10 @@ export type DiscussionListRowProps = {
 };
 
 function DiscussionListRow({ discussion, showRealNames, onClick }: DiscussionListRowProps) {
+	const { t } = useTranslation();
 	const formatDate = useTimeAgo();
 
-	const msg = normalizeThreadMessage(discussion);
+	const msg = normalizeThreadMessage(discussion, t);
 
 	const { name = discussion.u.username } = discussion.u;